+++ src/quadrature/gauleg.hh
@@ -4,22 +4,17 @@
  * \author Kent Budge
  * \date   Tue Sep 14 13:16:09 2004
  * \brief  Gauss-Legendre quadrature
- * \note   Copyright (C) 2016 Los Alamos National Security, LLC.
- *         All rights reserved.
- */
-//---------------------------------------------------------------------------//
-// $Id$
+ * \note   Copyright (C) 2016-2019 Triad National Security, LLC.
+ *         All rights reserved. */
 //---------------------------------------------------------------------------//
 
 #ifndef rtt_quadrature_gauleg_hh
 #define rtt_quadrature_gauleg_hh
 
-#include <limits>
-
-#include <gsl/gsl_sf_legendre.h>
-
 #include "ds++/Soft_Equivalence.hh"
 #include "units/PhysicalConstants.hh"
+#include <gsl/gsl_sf_legendre.h>
+#include <limits>
 
 namespace rtt_quadrature {
 //---------------------------------------------------------------------------//
@@ -49,14 +44,13 @@ namespace rtt_quadrature {
  * and provides the abscissas \f$ x_j\f$ and weights \f$ w_j \f$ for the
  * Gaussian formula provided above.
  * 
- * \arg \a FieldVector A random access container on a field type.
+ * \tparam FieldVector A random access container on a field type.
  *
  * \param x1 Start of integration interval
  * \param x2 End of integration interval
  * \param x On return, contains abscissae \f$x_j\f$ for quadrature.
  * \param w On return, contains weights \f$w_j\f$ for quadrature.
- * \param N Number of points in quadrature.
- * 
+ * \param n Number of points in quadrature. 
  */
 template <typename FieldVector>
 void gauleg(
@@ -115,7 +109,7 @@ void gauleg(
     } while (!soft_equiv(z, z1, tolerance));
 
     // Roots will be between -1 and 1.0 and symmetric about the origin.
-    int const idxSymPart(n - iroot - 1);
+    size_t const idxSymPart(n - iroot - 1);
 
     // Now, scale the root to tthe desired interval and put in its
     // symmetric counterpart.
